requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data processing
pandas==2.1.3
numpy==1.25.2
duckdb==0.9.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Notebook support
nbformat==5.9.2
jupyter==1.0.0

# LLM integration
langchain==0.2.10
langchain-openai==0.1.16
openai==1.40.6

# Web framework
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# Environment and config
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
pathlib2==2.3.7; python_version < "3.4"
typing-extensions==4.8.0